Updates/optimizations (#36)
* Refactor(validations): Replace User Validation with Zod Schemas

- Removed the old user validation logic using express-validator.
- Introduced new Zod schemas for user and project validation.
- Updated Prisma schema to include cascading delete behavior for user preferences and increased project description length.

* Refactor(auth): Reroute User Authentication Endpoints

- auth routes available at /api/v1/auth
- Added signup endpoint for user registration with password hashing and JWT generation.
- Implemented login endpoint to authenticate users and return a JWT upon successful login.
- Created logout endpoint to clear the authentication token and log users out.
- Integrated Zod validation for request bodies in signup and login processes.

* Refactor(auth): Enhance User Session Restoration

- Updated user session restoration to use findUniqueOrThrow for better error handling.
- Created ReqUserObject class to encapsulate user information (id and email).
- Added logging for user requests to track user activity.

* Refactor(accounts): Simplify User Management Functions

- Removed old authentication logic including signup, login, and logout functions.
- Introduced new user management functions: getAllUsers, getUserById, updateUserById, and deleteUserById.
- Implemented user validation using Zod schemas for update operations.
- Enhanced security by ensuring only the logged-in user can update or delete their account.
- Added comments for better code documentation and future implementation notes.

* Refactor(projects): Improve Project Management Functions

- Replaced inline Zod schema validation with a dedicated validation utility for better code organization.
- Added JSDoc comments for create, get, update, and delete project functions to enhance code documentation.
- Simplified project ID extraction from request parameters for clarity.
- Implemented user ownership checks to ensure only the project creator can update or delete their projects.
- Added a TODO comment for future implementation of filtering and pagination in the getAllProjects function.

+++ apps/backend/src/controllers/projectsControllers.js
@@ -1,22 +1,16 @@
 import { prisma } from "@repo/db";
-import { z } from "zod";
+import { validateRequestBody } from "../utils/validations/zod-error-formatter.js";
+import { projectSchema } from "../utils/validations/zod-schemas.js";
 
-const projectSchema = z.object({
-	title: z.string(),
-	description: z.string(),
-	githubUrl: z.string(),
-	isPublic: z.boolean().optional(),
-	thumbnailUrl: z.string().optional(),
-});
-
-// create a new project
+/**
+ * create a new project
+ * @param {Request} req
+ * @param {Response} res
+ * @param {Function} next
+ */
 export const createProject = async (req, res, next) => {
-	const { error } = projectSchema.safeParse(req.body);
-	if (error) {
-		next(error);
-		return;
-	}
 	try {
+		validateRequestBody(projectSchema, req, next);
 		const project = await prisma.project.create({
 			data: req.body,
 		});
@@ -26,7 +20,13 @@ export const createProject = async (req, res, next) => {
 	}
 };
 
-// get all projects
+// TODO: Add req.query params to handle filtering and pagination
+/**
+ * get all projects
+ * @param {Request} req
+ * @param {Response} res
+ * @param {Function} next
+ */
 export const getAllProjects = async (req, res, next) => {
 	try {
 		const projects = await prisma.project.findMany();
@@ -36,15 +36,20 @@ export const getAllProjects = async (req, res, next) => {
 	}
 };
 
-// get project by id
+/**
+ * get project by id
+ * @param {Request} req
+ * @param {Response} res
+ * @param {Function} next
+ */
 export const getProjectById = async (req, res, next) => {
 	try {
 		const { id } = req.params;
-		const projectId = id;
 
 		const project = await prisma.project.findUnique({
-			where: { id: projectId },
+			where: { id },
 		});
+
 		if (!project) {
 			return res.status(404).json({ error: "Project not found" });
 		}
@@ -54,37 +59,50 @@ export const getProjectById = async (req, res, next) => {
 	}
 };
 
-// update project
+/**
+ * update project
+ * @param {Request} req
+ * @param {Response} res
+ * @param {Function} next
+ */
 export const updateProjectById = async (req, res, next) => {
-	const { id } = req.params;
-	const projectId = id;
-	const { error } = projectSchema.safeParse(req.body);
-	if (error) {
-		next(error);
-		return;
-	}
+	const { id: projectId } = req.params;
+	const { id: userId } = req.user;
+
 	try {
+		validateRequestBody(projectSchema);
 		const existingProject = await prisma.project.findUnique({
 			where: { id: projectId },
 		});
+
 		if (!existingProject) {
 			return res.status(404).json({ error: "Project not found" });
 		}
+
+		// Only currently logged in user that created the project can update project
+		if (userId !== existingProject.ownerId) return res.status(403).json({ message: "Forbidden" });
+
 		const updatedProject = await prisma.project.update({
 			where: { id: projectId },
 			data: req.body,
 		});
+
 		res.json(updatedProject);
 	} catch (error) {
 		next(error);
 	}
 };
 
-// delete project
+/**
+ * delete project
+ * @param {Request} req
+ * @param {Response} res
+ * @param {Function} next
+ */
 export const deleteProjectById = async (req, res, next) => {
 	try {
-		const { id } = req.params;
-		const projectId = id;
+		const { id: projectId } = req.params;
+		const { id: userId } = req.user;
 
 		const existingProject = await prisma.project.findUnique({
 			where: { id: projectId },
@@ -92,6 +110,9 @@ export const deleteProjectById = async (req, res, next) => {
 		if (!existingProject) {
 			return res.status(404).json({ error: "Project not found" });
 		}
+
+		if (userId !== existingProject.ownerId) return res.status(403).json({ message: "Forbidden" });
+
 		await prisma.project.delete({
 			where: { id: projectId },
 		});
